requirements.txt
dataclasses==0.7
influxdb==5.2.3
jsons==1.1.0
Flask==1.1.1
flask-accepts==0.15.5
flask-cors==3.0.8
flask-restplus==0.13.0
marshmallow==3.3.0
marshmallow-dataclass==7.2.1
python-dateutil==2.8.1
gunicorn==20.0.4
pytz==2019.3

# Required to pin because werkzeug 1.0 does not work with flask-restplus.
werkzeug==0.16.1

# There's a bug in the current latest release (0.5.3) with the newer version of the
# attrs lib. So instead we hardcode the GIT commit hash of a working version.
attrs==19.2.0
git+https://github.com/weaveworks/grafanalib#b0375148133e1f9b0c78777c8ec818cf018235e0

# Error reporting is not very good in the main line. There is a merge request but
# for now use our own.
grafana-api==0.9.2

# Development dependencies.
ptvsd
watchdog==0.9.0
